# release_candidates/rc0/serverside/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Api do Smart Parking da Mauá"
  version: "1.0.0"
  title: "Smart Parking"
host: "localhost:8080"
basePath: "/v1"
...
schemes:
- "http"
paths:
...
  /estacionamentos:
    get:
      tags:
      - "front"
      summary: "Retorna os estacionamentos e seus id's"
      description: "Retorna um array com o nome e id de todos os estacionamentos registrados"
      operationId: "estacionamentosGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EstacionamentoResponse"
        400:
          description: "Invalid status value"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Front"
  /estacionamentos/{id}/findByHour:
    get:
      tags:
      - "front"
      summary: "Retorna a quantidade de carros por hora"
      description: "Retorna dois arrays de 12 elementos cada, sendo um de entrada\
        \ e um de saida, com a quantidade de carros no estacionamento a cada 5 minutos,\
        \ completando uma hora no total"
      operationId: "estacionamentosIdFindByHourGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id do estacionamento"
        required: true
        type: "integer"
        format: "int32"
      - name: "timestamp"
        in: "query"
        description: "Data desejado"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HourResponse"
        400:
          description: "Invalid status value"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Front"
  /estacionamentos/{id}/findByDay:
    get:
      tags:
      - "front"
      summary: "Retorna a quantidade de carros por dia"
      description: "Retorna dois arrays de 24 elementos cada, sendo um de entrada\
        \ e um de saida, com a quantidade de carros no estacionamento a cada hora,\
        \ completando um dia no total"
      operationId: "estacionamentosIdFindByDayGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id do estacionamento"
        required: true
        type: "integer"
        format: "int32"
      - name: "timestamp"
        in: "query"
        description: "Data desejado"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DayResponse"
        400:
          description: "Invalid status value"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Front"
  /estacionamentos/{id}/findByWeek:
    get:
      tags:
      - "front"
      summary: "Retorna a quantidade de carros dos proximos 7 dias"
      description: "Retorna 7 objetos, um para cada dia da semana, e cada objeto possui\
        \ dois arrays de 24 elementos cada, sendo um de entrada e um de saida, com\
        \ a quantidade de carros no estacionamento a cada hora, completando um dia\
        \ no total, para cada dia da semana"
      operationId: "estacionamentosIdFindByWeekGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id do estacionamento"
        required: true
        type: "integer"
        format: "int32"
      - name: "timestamp"
        in: "query"
        description: "Data desejado"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/WeekResponse"
        400:
          description: "Invalid status value"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Front"
  /estacionamentos/{id}/findByMonth:
    get:
      tags:
      - "front"
      summary: "Retorna a quantidade de carros por mes"
      description: "Retorna 5 objetos, um para cada semana, e cada objeto possui 7\
        \ objetos, um para cada dia da semana, e cada objeto possui dois arrays de\
        \ 24 elementos cada, sendo um de entrada e um de saida, com a quantidade de\
        \ carros no estacionamento a cada hora, completando um dia no total, para\
        \ cada dia da semana"
      operationId: "estacionamentosIdFindByMonthGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id do estacionamento"
        required: true
        type: "integer"
        format: "int32"
      - name: "timestamp"
        in: "query"
        description: "Data desejado"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/MonthResponse"
        400:
          description: "Invalid status value"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Front"
definitions:
...
  Estacionamento:
    type: "object"
    properties:
      name:
        type: "string"
      id:
        type: "integer"
        format: "int32"
  EstacionamentoResponse:
    type: "object"
    properties:
      estacionamentos:
        type: "array"
        example:
        - name: "Campus"
          id: 0
        - name: "Portão Principal"
          id: 1
        - name: "Divisão de Motores"
          id: 2
        - name: "Bolsão H"
          id: 3
        - name: "Bolsão U"
          id: 4
        items:
          $ref: "#/definitions/Estacionamento"
    example:
      estacionamentos:
      - name: "Campus"
        id: 0
      - name: "Portão Principal"
        id: 1
      - name: "Divisão de Motores"
        id: 2
      - name: "Bolsão H"
        id: 3
      - name: "Bolsão U"
        id: 4
  HourResponse:
    type: "object"
    properties:
      entrada:
        type: "array"
        example:
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        items:
          type: "integer"
          format: "int32"
          maximum: 12
      saida:
        type: "array"
        example:
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        items:
          type: "integer"
          format: "int32"
          maximum: 12
    example:
      entrada:
      - 10
      - 13
      - 17
      - 15
      - 8
      - 8
      - 9
      - 14
      - 16
      - 18
      - 20
      - 21
      saida:
      - 10
      - 13
      - 17
      - 15
      - 8
      - 8
      - 9
      - 14
      - 16
      - 18
      - 20
      - 21
  DayResponse:
    type: "object"
    properties:
      entrada:
        type: "array"
        example:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        items:
          type: "integer"
          format: "int32"
          maximum: 24
      saida:
        type: "array"
        example:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        items:
          type: "integer"
          format: "int32"
          maximum: 24
    example:
      entrada:
      - 1
      - 0
      - 0
      - 0
      - 0
      - 0
      - 10
      - 13
      - 17
      - 15
      - 8
      - 8
      - 9
      - 14
      - 16
      - 18
      - 20
      - 21
      - 17
      - 15
      - 12
      - 7
      - 3
      - 2
      saida:
      - 1
      - 0
      - 0
      - 0
      - 0
      - 0
      - 10
      - 13
      - 17
      - 15
      - 8
      - 8
      - 9
      - 14
      - 16
      - 18
      - 20
      - 21
      - 17
      - 15
      - 12
      - 7
      - 3
      - 2
  WeekResponse:
    type: "object"
    properties:
      segunda:
        $ref: "#/definitions/DayResponse"
      terca:
        $ref: "#/definitions/DayResponse"
      quarta:
        $ref: "#/definitions/DayResponse"
      quinta:
        $ref: "#/definitions/DayResponse"
      sexta:
        $ref: "#/definitions/DayResponse"
      sabado:
        $ref: "#/definitions/DayResponse"
      domingo:
        $ref: "#/definitions/DayResponse"
    example:
      quarta:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
      segunda:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
      quinta:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
      domingo:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
      sexta:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
      sabado:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
      terca:
        entrada:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
        saida:
        - 1
        - 0
        - 0
        - 0
        - 0
        - 0
        - 10
        - 13
        - 17
        - 15
        - 8
        - 8
        - 9
        - 14
        - 16
        - 18
        - 20
        - 21
        - 17
        - 15
        - 12
        - 7
        - 3
        - 2
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"